backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-Bcrypt==1.0.1
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
PyJWT==2.8.0
python-dotenv==1.0.0
psycopg2-binary==2.9.7
python-dateutil==2.8.2
cloudinary==1.36.0
gunicorn==21.2.0